split device logic into shared lib
